requirements.txt
asgiref==3.7.2
Django==4.2.8
django-cleanup==8.0.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pep8==1.7.1
Pillow==10.1.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
